Basic Structure
Basic structure to handle sensors as well as some higher-level planning
regarding navigation and states. There is a "Sensor" abstract class for
deriving new sensors. New sensors should be added to the SensorHandler.
Put an array in for Obstacle Detection...I believe we should average a
sample of X values to determine if we are approaching an obstacle.
Sensors are unreliable, need multiple samples. Held off on many
implementations until we decide what we want to reuse from last year.